+++ packages/swirl-components/src/components/swirl-shell-layout/swirl-shell-layout.stories.ts
@@ -65,11 +65,13 @@ const Template = (args) => {
         </swirl-shell-navigation-item>
       </swirl-stack>
     </div>
+    <swirl-banner slot="banner" action-label="Update" content="New version available." dismissable="false" intent="info" size="s"></swirl-banner>
     <div slot="main">
       <swirl-app-layout
         app-name="Chat"
         navigation-label="Items"
       >
+        <swirl-banner slot="banner" action-label="Details" content="On leave until 16.05.2023" dismissable="false" intent="info"></swirl-banner>
         <swirl-button
           hide-label
           icon="<swirl-icon-menu></swirl-icon-menu>"