+++ packages/swirl-components/src/components/swirl-shell-layout/swirl-shell-layout.tsx
@@ -7,6 +7,7 @@ import { getDesktopMediaQuery, getActiveElement } from "../../utils";
  * @slot logo-collapsed - Logo shown inside collapsed sidebar.
  * @slot tools - Items shown in the upper sidebar part.
  * @slot main-navigation - Items shown in the lower sidebar part.
+ * @slot banner - Used to show a swirl-banner on top of the page.
  * @slot main - Contents of the main area.
  */
 @Component({
@@ -93,6 +94,7 @@ export class SwirlShellLayout {
 
   private onSidebarClick = () => {
     if (this.collapsedSidebar) {
+      (document.activeElement as HTMLElement)?.blur();
       (getActiveElement() as HTMLElement)?.blur();
     }
   };
@@ -125,6 +127,9 @@ export class SwirlShellLayout {
     return (
       <Host>
         <div class={className}>
+          <div class="shell-layout__banner">
+            <slot name="banner"></slot>
+          </div>
           <div
             class={sidebarWrapperClassName}
             onClick={this.onSidebarClick}